requirements.txt
flask
google-generativeai
sentence-transformers
faiss-cpu
datasets
python-dotenv
pyarrow
numpy<2.0